\name{TopSpace-package}
\alias{TopSpace-package}
\alias{TopSpace}
\docType{package}
\title{
\packageTitle{TopSpace}
}
\description{
\packageDescription{TopSpace}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{TopSpace}

Maintainer: \packageMaintainer{TopSpace}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = TopSpace)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{TopSpace}}
%\section{Documentation Index}{\packageIndices{TopSpace}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
