% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_mcmc_TopSpace.R
\name{init.mcmc.TopSpace}
\alias{init.mcmc.TopSpace}
\title{Initialize the MCMC algorithm for TopSpace}
\usage{
init.mcmc.TopSpace(
  X,
  nbd.sizes,
  H,
  K,
  coord,
  kernel.params = c(0.01, 1),
  poly.degree = 20L
)
}
\arguments{
\item{X}{}

\item{nbd.sizes}{}

\item{H}{}

\item{K}{}

\item{coord}{}

\item{kernel.params}{}

\item{poly.degree}{}
}
\description{
Initialize the MCMC algorithm for TopSpace
}
