% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_TopSpace.R
\name{mcmc.TopSpace}
\alias{mcmc.TopSpace}
\title{An efficient implementation of Markov chain Monte Carlo for TopSpace}
\usage{
mcmc.TopSpace(
  X,
  nbd.sizes,
  H,
  K,
  init,
  priors,
  n.mcmc,
  n.burnin,
  step.show = 100,
  learning.rate = 0.01,
  decay = 0.01,
  n.leapfrog = 10,
  prop.subsample = 0.5,
  thres.merge = 0.01
)
}
\arguments{
\item{X}{}

\item{nbd.sizes}{}

\item{H}{}

\item{K}{}

\item{init}{}

\item{priors}{}

\item{n.mcmc}{}

\item{n.burnin}{}

\item{step.show}{}

\item{learning.rate}{}

\item{decay}{}

\item{n.leapfrog}{}

\item{prop.subsample}{}

\item{thres.merge}{}
}
\description{
An efficient implementation of Markov chain Monte Carlo for TopSpace
}
