% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSCLC.R
\docType{data}
\name{NSCLC}
\alias{NSCLC}
\title{Non-small Cell Lung Cancer Dataset}
\format{
A list 
\describe{
  \item{variable1}{Description of variable1}
  \item{variable2}{Description of variable2}
  \item{variable3}{Description of variable3}
}
}
\usage{
NSCLC
}
\description{
A dataset of multiplex immunohistochemistry images of tissue sections from 153 patients with non-small cell lung cancer.
}
\keyword{datasets}
